// apps/tourism/platform/src/types/hotel.ts
export interface definitions {
  AdminActiveHotelBody: {
    active?: boolean;
  };
  AdminActiveSpecialContentBody: {
    active?: boolean;
  };
  AdminAddSpecialContentBody: {
    title?: string;
    content?: string;
    url?: string;
    active?: boolean;
  };
// ...
  AdminUpdatePackageBody: {
    name?: string;
    description?: string;
    adjustmentType?: string;
    /** Format: float */
    adjustmentValue?: number;
    isActive?: boolean;
    nameEng?: string;
    rules?: string;
    rulesEng?: string;
  };
// ...
  HotelAddPassengerV2Body: {
    rooms?: definitions["hotelPassengerV2"][];
  };
// ...
  apihotelRoomPrice: {
    roomId?: string;
    /** only room price */
    price?: definitions["hotelPriceDetail"];
    halfCharge?: definitions["apihotelHalfDayPrice"];
    extraBed?: definitions["hotelExtraBed"];
    refund?: definitions["hotelRefundPenalty"];
    /**
     * all the extra prices with room price
     * Format: int64
     */
    totalPrice?: number;
  };
// ...
  hotelAdminIntOrder: {
    orderId?: string;
    status?: definitions["hotelOrderStatus"];
    hotelInfo?: definitions["hotelHotelDetails"];
    orderNumber?: string;
    room?: definitions["hotelAdminIntRoom"][];
// ...
  };
// ...
  hotelExchange: {
    baseExchange?: definitions["hotelExchangePrice"];
    totalExchange?: definitions["hotelExchangePrice"];
    /** Format: double */
    unit?: number;
    /** Format: double */
    rate?: number;
  };
// ...
  rpcStatus: {
    /** Format: int32 */
    code?: number;
    message?: string;
    details?: definitions["protobufAny"][];
  };
}